reduced frame rate for animation
